+++ CHANGELOG.md
@@ -35,6 +35,7 @@ Evolutions :
 * Ajout du grade dans la génération des PV et ODJ pour les préventionnistes attachés au dossier
 * Le bloc des établissements sans prochaines visites périodiques programmées tient maintenant compte les dossiers de type avant ouverture
 * Ajout de la possibilité de replier les blocs du tableau de bord
+* Ajout d'un cache APC sur les métadonnées de l'ORM hors DEV
 
 Corrections :
 * Impossibilité de créer des dossiers en décembre : mauvais validateur de date